requirements.txt
pandas==0.25.3
numpy==1.17.4
matplotlib==3.1.1
opencv-python
Pillow==6.2.1
scikit_learn==0.21.3
scikit-image
tqdm==4.39.0